refactor: Moved color-picker functionality to a service
